requirements.txt
google-generativeai
pillow
streamlit
gspread
oauth2client